# client/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eyened API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/import/image": {
      "post": {
        "summary": "Import Single Image",
        "operationId": "import_single_image_import_image_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import/run_cfi_models": {
      "post": {
        "summary": "Enqueue Run Cfi Models",
        "description": "Queue one RQ job per CFI attribute model (``cfi-roi``, ``cfi-quality``, ...).",
        "operationId": "enqueue_run_cfi_models_import_run_cfi_models_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCfiModelsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import/run_cfi_amd": {
      "post": {
        "summary": "Enqueue Run Cfi Amd",
        "operationId": "enqueue_run_cfi_amd_import_run_cfi_amd_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCfiAmdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import/run_layer_segmentation": {
      "post": {
        "summary": "Enqueue Run Layer Segmentation",
        "operationId": "enqueue_run_layer_segmentation_import_run_layer_segmentation_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunImageIdsJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import/update_thumbnails": {
      "post": {
        "summary": "Post Import Update Thumbnails",
        "description": "Queue a job that scans the DB for missing thumbnails (CLI-equivalent).",
        "operationId": "post_import_update_thumbnails_import_update_thumbnails_post",
        "parameters": [
          {
            "name": "failed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include images with empty ThumbnailPath (failed previous run), like ``eorm update-thumbnails --failed``.",
              "default": false,
              "title": "Failed"
            },
            "description": "Include images with empty ThumbnailPath (failed previous run), like ``eorm update-thumbnails --failed``."
          },
          {
            "name": "print_errors",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Print per-image errors on the worker, like ``eorm update-thumbnails --print-errors``.",
              "default": false,
              "title": "Print Errors"
            },
            "description": "Print per-image errors on the worker, like ``eorm update-thumbnails --print-errors``."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import/update_thumbnails_for_image_ids": {
      "post": {
        "summary": "Post Import Update Thumbnails For Image Ids",
        "description": "Queue thumbnail generation for the given ``image_ids`` only.",
        "operationId": "post_import_update_thumbnails_for_image_ids_import_update_thumbnails_for_image_ids_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThumbnailsForImageIdsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import/status/{task_id}": {
      "get": {
        "summary": "Get Status",
        "description": "Return a queued job's status and result.\n\nAuthenticated, but not scope-aware, and that is not an oversight: every RQ\nentrypoint returns a bare bool (pinned by\ntest_every_rq_entrypoint_returns_a_bare_bool), so there is no project data\nin the response to filter. Job ids are uuid4 and not enumerable here, so\nthis is a capability control -- but a capability that leaks through a log\nline or a screenshot would otherwise need no credential at all.",
        "operationId": "get_status_import_status__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportOptions": {
        "properties": {
          "include_stack_trace": {
            "type": "boolean",
            "title": "Include Stack Trace",
            "description": "If True, include stack trace in the error response",
            "default": false
          }
        },
        "type": "object",
        "title": "ImportOptions",
        "description": "Options for the import process"
      },
      "ImportRequest": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ImportRow"
          },
          "options": {
            "$ref": "#/components/schemas/ImportOptions"
          }
        },
        "type": "object",
        "required": [
          "data",
          "options"
        ],
        "title": "ImportRequest"
      },
      "ImportResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stack_trace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Trace"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ImportResponse"
      },
...
      "RunCfiAmdRequest": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Image Ids"
          },
          "overwrite": {
            "type": "boolean",
            "title": "Overwrite",
            "default": false
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size",
            "default": 8
          },
          "n_workers": {
            "type": "integer",
            "title": "N Workers",
            "default": 12
          }
        },
        "type": "object",
        "required": [
          "image_ids"
        ],
        "title": "RunCfiAmdRequest"
      },
      "RunCfiModelsRequest": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Image Ids",
            "description": "ImageInstance IDs to process (e.g. ColorFundus IDs from a completed import)."
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "cfi-roi",
                  "cfi-keypoints",
                  "cfi-odfd",
                  "cfi-quality"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Single CFI attribute pipeline, or omit to run all attribute models."
          },
          "overwrite": {
            "type": "boolean",
            "title": "Overwrite",
            "description": "If true, re-run even when results already exist.",
            "default": false
          },
          "commit_interval": {
            "type": "integer",
            "title": "Commit Interval",
            "default": 100
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size",
            "default": 8
          },
          "n_workers": {
            "type": "integer",
            "title": "N Workers",
            "default": 16
          }
        },
        "type": "object",
        "required": [
          "image_ids"
        ],
        "title": "RunCfiModelsRequest",
        "description": "Queue a model job for the given image instance IDs (CFI attributes or segmentation)."
      },
      "RunImageIdsJobRequest": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Image Ids"
          },
          "overwrite": {
            "type": "boolean",
            "title": "Overwrite",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "image_ids"
        ],
        "title": "RunImageIdsJobRequest"
      },
...
      "TaskResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "task_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Ids"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "TaskResponse"
      },
...
      "UpdateThumbnailsForImageIdsRequest": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Image Ids",
            "description": "ImageInstance IDs to generate thumbnails for."
          },
          "print_errors": {
            "type": "boolean",
            "title": "Print Errors",
            "description": "Log per-image errors to worker stdout.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "image_ids"
        ],
        "title": "UpdateThumbnailsForImageIdsRequest",
        "description": "Queue thumbnail generation for specific instances (same processing as bulk update)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}